Add unit test for server cache
